requirements.txt
streamlit
llama-index
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-vector-stores-postgres  # <-- This is the corrected line
llama-index-readers-file            # <-- Added for robustness
psycopg2-binary
SQLAlchemy
python-dotenv